+++ lib/getch/gentoo/sources.rb
@@ -43,7 +43,7 @@ module Getch
 
         puts "Adding support for cryptsetup."
         bask("-a cryptsetup")
-        exec("euse -E cryptsetup") if !Helpers::grep?(make_conf, /cryptsetup/)
+        Getch::Chroot.new("euse -E cryptsetup").run! unless Helpers::grep?(make_conf, /cryptsetup/)
         Getch::Emerge.new("sys-fs/cryptsetup").pkg!
       end